Use ep addr hash list in RDMA protocol
This patch changes the plugin's endpoint creation behavior (in
ep-per-comm mode) as follows:
1. All send communicators will use the same endpoint.
2. When creating an endpoint for receive communicators, first check if
any existing endpoint is not already connected to the source endpoint.
Only create a new recv endpoint if all existing endpoints are already
connected to the source endpoint.

Currently, the endpoint list is stored with the device structure and not
locked, so **this feature is not safe to use with threads**. It will
work correctly in single-threaded applications.

This feature is controlled by a parameter,
`OFI_NCCL_ENDPOINT_PER_UNIQUE_SRC` (default 0, 1 enables the feature).

Signed-off-by: Eric Raut <eraut@amazon.com>