rdma: Ignore Libfabric canceled events
In preparation for sharing a completion queue. We close endpoints while receive
operations are still in progress, which causes these to be canceled and
appear in the error queue. Today, we close the cq immediately so don't
see these events. However, future work will share the cq between
multiple endpoints, so we need to explicitly handle (and ignore) these
events.

Signed-off-by: Eric Raut <eraut@amazon.com>