rdma: use ep_addr_list for endpoint-per-comm code
This patch changes the plugin's endpoint creation behavior (in
ep-per-comm mode) as follows:
1. All send communicators will use the same endpoint.
2. When creating an endpoint for receive communicators, first check if
any existing endpoint is not already connected to the source endpoint.
Only create a new recv endpoint if all existing endpoints are already
connected to the source endpoint.

This feature is used when `OFI_NCCL_ENDPOINT_PER_COMM=1`.

Signed-off-by: Eric Raut <eraut@amazon.com>