Update to a more sensible public interface
Makes the public interface (nccl_ofi_cm.h) leaner, with clear
distinction of which methods (public) should be used by the caller
(transport) and which methods are private implementation details.

Moved some functionality that will become separate internal objects for
better separation of concerns.

Implementation is not done yet, so this definitely does not compile.